refactor statistics calculation
